# docs/openspec/registry.openapi.yaml
openapi: 3.0.3
info:
  title: Virbius Control Plane API (MVP)
  version: 1.3.0-mvp
  description: |
    **virbius-control** 单 Java 进程（Registry + Admin 合并）。

    - **规范面**：`/api/v1/admin/tenants/{tenantId}/...`，响应 `{ code, message, data, timestamp }`（`code=0` 成功）。
    - **兼容面**：`/api/v1/tenants/{tenantId}/...`（Legacy，裸 JSON），逐步废弃。
    - 规则真源：`rule_history`；`runtime` 见 DESIGN §8.5.0。
    - **规则级 enforce / canary**（管侧真拦 + 云 PolicyMerge）：见 [rule-level-enforce.md](./rule-level-enforce.md)。
    - PoC 发布为 **同步 HTTP 200**（非 202 异步）。
    - Bundle YAML 导入（`draft/import`）**未实现**，Post-MVP。
servers:
  - url: http://virbius-control:8080
    description: PoC（Registry + Admin UI 同端口）
...
paths:
...
  /api/v1/admin/tenants/{tenantId}/bundles:
    get:
      tags: [admin-bundles]
      summary: 列出 Bundle
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultBundleSummaryList'
    post:
      tags: [admin-bundles]
      summary: 创建 Bundle（draft）
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBundleRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultBundleSummary'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}:
    get:
      tags: [admin-bundles]
      summary: 获取 Bundle 详情
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultBundleDetail'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}:
    get:
      tags: [admin-bundles]
      summary: 获取 Bundle 版本
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultBundleVersion'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/metadata:
    get:
      tags: [admin-bundles]
      summary: 获取 Bundle 元数据（含 context_bindings）
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultBundleMetadata'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/metadata/context-bindings:
    put:
      tags: [admin-bundles]
      summary: 替换逻辑变量映射（context_bindings）
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
        - name: sync
          in: query
          schema:
            type: boolean
            default: true
          description: true 时同步 gw/cloud 名单规则并写 gateway 产物
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextBindingsRequest'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultContextBindingsUpdate'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/metadata/sync-gateway:
    post:
      tags: [admin-bundles]
      summary: 同步 gateway 名单产物（不写 cloud）
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/publish:
    post:
      tags: [admin-publish]
      summary: 发布 Bundle 版本（同步执行状态机）
      description: |
        PoC：同步完成并返回最终 status（通常 active）。
        draft → validating → eval_running → compiling → syncing → active
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: 发布完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultPublishStatus'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/status:
    get:
      tags: [admin-publish]
      summary: 查询发布状态与 sync_ack
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultPublishStatus'
  /api/v1/admin/tenants/{tenantId}/access-lists:
    get:
      tags: [admin-access-lists]
      summary: 查询全部名单
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/{dimension}/{polarity}:
    get:
      tags: [admin-access-lists]
      summary: 查询单维名单
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
    put:
      tags: [admin-access-lists]
      summary: 整表替换名单并同步产物
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessListEntriesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries:
    post:
      tags: [admin-access-lists]
      summary: 追加名单条目
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessListEntriesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries/{value}:
    delete:
      tags: [admin-access-lists]
      summary: 删除名单条目
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
        - name: value
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/sync-rules:
    post:
      tags: [admin-access-lists]
      summary: 同步名单 → 规则行 + gateway/edge 产物
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/push-engine:
    post:
      tags: [admin-access-lists]
      summary: 推送 cloud 名单至 engine RuleCache
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
  /api/v1/admin/tenants/{tenantId}/access-lists/sync-and-publish:
    post:
      tags: [admin-access-lists]
      summary: 同步名单后发布 Bundle
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - name: bundleId
          in: query
          schema:
            type: string
            default: poc-default
        - name: version
          in: query
          schema:
            type: string
            default: "0.1.0"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResultObject'
...
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取 Bundle 版本"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/metadata:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 获取 Bundle 元数据（无 PUT context-bindings）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleMetadata'
  /api/v1/tenants/{tenantId}/bundles/{bundleId}/versions/{version}/publish:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 发布 Bundle（同步 200）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/bundleId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: 发布完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishStatus'
...
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}:
    get:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 查询单维名单（无 PUT 整表替换）"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 追加名单条目"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessListEntriesRequest'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/{dimension}/{polarity}/entries/{value}:
    delete:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 删除名单条目"
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/dimension'
        - $ref: '#/components/parameters/polarity'
        - name: value
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/sync-rules:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 同步名单规则"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
  /api/v1/tenants/{tenantId}/access-lists/push-engine:
    post:
      deprecated: true
      tags: [legacy]
      summary: "[Legacy] 推送 engine"
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: OK
...
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        example: default
    bundleId:
      name: bundleId
      in: path
      required: true
      schema:
        type: string
        example: poc-default
    version:
      name: version
      in: path
      required: true
      schema:
        type: string
        example: "0.1.0"
...
    dimension:
      name: dimension
      in: path
      required: true
      schema:
        type: string
        enum: [keyword, user_id, device_id, ip_cidr, var]
    polarity:
      name: polarity
      in: path
      required: true
      schema:
        type: string
        enum: [deny, allow]
  responses:
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResultError'
  schemas:
...
    ApiResultError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    ApiResultObject:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    ApiResultBundleSummaryList:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/BundleSummary'
    ApiResultBundleSummary:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/BundleSummary'
    ApiResultBundleDetail:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/BundleDetail'
    ApiResultBundleVersion:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    ApiResultBundleMetadata:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/BundleMetadata'
    ApiResultContextBindingsUpdate:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            updated:
              type: boolean
            context_bindings:
              $ref: '#/components/schemas/ContextBindingsBlock'
            context_vars:
              type: array
              items:
                $ref: '#/components/schemas/ContextVarBindingRow'
            sync:
              type: object
              additionalProperties: true
    ApiResultPublishStatus:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/PublishStatus'
...
    BundleSummary:
      type: object
      properties:
        bundle_id:
          type: string
        tenant_id:
          type: string
        latest_version:
          type: string
        status:
          type: string
          enum: [draft, validating, eval_running, compiling, syncing, active, failed]
    CreateBundleRequest:
      type: object
      required: [bundle_id]
      properties:
        bundle_id:
          type: string
    BundleDetail:
      allOf:
        - $ref: '#/components/schemas/BundleSummary'
        - type: object
          properties:
            rules:
              type: array
              items:
                $ref: '#/components/schemas/RulePointer'
    BundleMetadata:
      type: object
      properties:
        tenant_id:
          type: string
        bundle_id:
          type: string
        version:
          type: string
        status:
          type: string
        metadata:
          type: object
          additionalProperties: true
        context_bindings:
          $ref: '#/components/schemas/ContextBindingsBlock'
        context_vars:
          type: array
          items:
            $ref: '#/components/schemas/ContextVarBindingRow'
    ContextBindingsRequest:
      type: object
      required: [vars]
      properties:
        vars:
          type: array
          items:
            $ref: '#/components/schemas/ContextVarBindingRow'
    ContextVarBindingRow:
      type: object
      required: [logical, from]
      properties:
        logical:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        from:
          type: string
          enum: [query, header, subject, network]
        name:
          type: string
          description: from=query|header 时 HTTP 名；缺省=logical
        field:
          type: string
          description: from=subject|network 时字段名；缺省=logical
    ContextBindingsBlock:
      type: object
      properties:
        version:
          type: integer
          example: 1
        vars:
          type: object
          additionalProperties:
            type: object
            properties:
              from:
                type: string
                enum: [query, header, subject, network]
              name:
                type: string
              field:
                type: string
    AccessListEntriesRequest:
      type: object
      properties:
        value:
          type: string
        values:
          type: array
          items:
            type: string
    PublishStatus:
      type: object
      properties:
        bundle_id:
          type: string
        bundle_version:
          type: string
        status:
          type: string
        publish_id:
          type: string
        sync_ack:
          type: object
          additionalProperties: true
    RulePointer:
      type: object
      properties:
        rule_id:
          type: string
        current_revision:
          type: integer
        layer:
          type: string
        runtime:
          type: string
          enum: [lua-dsl, lua, prompt, groovy, list_match, cumulative]
        reason_code:
          type: string
        intent_action:
          type: string
          enum: [allow, deny, captcha, review]
          default: deny
          description: 规则命中后的希望动作；见 rule-level-enforce.md
        enforce_mode:
          type: string
          enum: [dry_run, canary, full]
          description: 规则级执行模式；见 rule-level-enforce.md
        canary_percent:
          type: integer
          minimum: 1
          maximum: 100
          nullable: true
          description: enforce_mode=canary 时必填；full/dry_run 时应省略或 null
        rule_status:
          type: string
          enum: [draft, active, disabled]
          description: |
            draft=配置中不进执行面；active=参与匹配；disabled=停用不可编辑。
            流转：draft→active|disabled；active→disabled；disabled→draft。
...